+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Ammo Count HUD (OpenMW)
 
+## 1.1
+
+- HUD no longer can be positioned outside of your screen
+- Added text shadow setting
+- Minor optimizations
+- Slight changes to the settings names
+
 ## 1.0.1
 
 - Minor performance improvements